+++ src/Controller/SecurityController.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * @copyright Zicht Online <https://zicht.nl>
+ */
+
+namespace Zicht\Bundle\AdminBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
+
+class SecurityController extends AbstractController
+{
+    /** {@inheritDoc} */
+    public static function getSubscribedServices()
+    {
+        return array_merge(parent::getSubscribedServices(), [AuthenticationUtils::class => AuthenticationUtils::class]);
+    }
+
+    /**
+     * The login handling, rendering of the login form etc
+     *
+     * @return Response
+     */
+    public function loginAction()
+    {
+        $authenticationUtils = $this->get('security.authentication_utils');
+
+        // get the login error if there is one
+        $error = $authenticationUtils->getLastAuthenticationError();
+
+        // last username entered by the user
+        $lastUsername = $authenticationUtils->getLastUsername();
+
+        return $this->render(
+            'ZichtAdminBundle:Security:login.html.twig',
+            [
+                // last username entered by the user
+                'last_username' => $lastUsername,
+                'error' => $error,
+            ]
+        );
+    }
+}