+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added|Changed|Deprecated|Removed|Fixed|Security
 
+## 6.0.0 - 2020-05-15
+### Added
+- Support for Symfony 4.x
+### Removed
+- Support for Symfony 3.x
+### Fixed
+- Fixed custom sorting issue in Tree Admin caused by sonata-project/doctrine-orm-admin-bundle 3.15
+### Changed
+- Removed Zicht(Test)/Bundle/AdminBundle/ directory depth: moved all code up directly into src/ and test/
+
 ## 5.2.0 - 2020-05-14
 ### Fixed
 - Fixed composer.json (psr-4 autoloader, dependencies), fixed linter errors, fixed tests